# Dev: upgrade prettier to 2.0 and switch to prettyquick and reduce prettify time 96%.