--- a/charts/scatterCharts/PointsWithLabels.tsx
+++ b/charts/scatterCharts/PointsWithLabels.tsx
@@ -287,9 +287,7 @@ export class PointsWithLabels extends React.Component<PointsWithLabelsProps> {
     }
 
     @computed private get fontScale(): (d: number) => number {
-        return scaleLinear()
-            .range([10, 13])
-            .domain(this.sizeScale.domain())
+        return scaleLinear().range([10, 13]).domain(this.sizeScale.domain())
     }
 
     @computed private get labelFontFamily(): string {
@@ -874,7 +872,7 @@ export class PointsWithLabels extends React.Component<PointsWithLabelsProps> {
         this.animSelection = select(this.base.current).selectAll("circle")
 
         this.animSelection
-            .each(function() {
+            .each(function () {
                 const circle = this as SVGCircleElement
                 radiuses.push(circle.getAttribute("r") as string)
                 circle.setAttribute("r", "0")