--- a/db/knexMigrations/20190305234711_DenormalizeLatestCountryData.ts
+++ b/db/knexMigrations/20190305234711_DenormalizeLatestCountryData.ts
@@ -1,6 +1,6 @@
 import * as Knex from "knex"
 
-exports.up = async function(knex: Knex): Promise<any> {
+exports.up = async function (knex: Knex): Promise<any> {
     await knex.schema.createTable("country_latest_data", t => {
         t.string("country_code")
         t.integer("variable_id").references("variables.id")
@@ -10,6 +10,6 @@ exports.up = async function(knex: Knex): Promise<any> {
     })
 }
 
-exports.down = async function(knex: Knex): Promise<any> {
+exports.down = async function (knex: Knex): Promise<any> {
     await knex.schema.dropTable("country_latest_data")
 }