--- a/db/model/ChartRevision.ts
+++ b/db/model/ChartRevision.ts
@@ -18,15 +18,9 @@ export class ChartRevision extends BaseEntity {
     @Column() createdAt!: Date
     @Column() updatedAt!: Date
 
-    @ManyToOne(
-        () => User,
-        user => user.editedCharts
-    )
+    @ManyToOne(() => User, user => user.editedCharts)
     user!: User
 
-    @ManyToOne(
-        () => Chart,
-        chart => chart.logs
-    )
+    @ManyToOne(() => Chart, chart => chart.logs)
     chart!: Chart
 }