--- a/db/model/Variable.ts
+++ b/db/model/Variable.ts
@@ -113,12 +113,14 @@ export async function writeVariableCSV(
         [variableIds]
     )
 
-    const dataQuery: Promise<{
-        variableId: number
-        entity: string
-        year: number
-        value: string
-    }[]> = db.query(
+    const dataQuery: Promise<
+        {
+            variableId: number
+            entity: string
+            year: number
+            value: string
+        }[]
+    > = db.query(
         `
         SELECT
             data_values.variableId AS variableId,