--- a/explorer/admin/ExplorerBaker.tsx
+++ b/explorer/admin/ExplorerBaker.tsx
@@ -183,9 +183,7 @@ const ExplorerPage = (props: ExplorerPageSettings) => {
             subnavId={subnavId}
             subnavCurrentId={subnavCurrentId}
         />
-    ) : (
-        undefined
-    )
+    ) : undefined
 
     return (
         <html>