--- a/site/client/covid/CovidTimeSeriesValue.tsx
+++ b/site/client/covid/CovidTimeSeriesValue.tsx
@@ -42,8 +42,6 @@ export const CovidTimeSeriesValue = ({
                     {formattedDate ?? formatDate(date)}
                 </span>
             </>
-        ) : (
-            undefined
-        )}
+        ) : undefined}
     </div>
 )