Instead of having prettify:check reexamine the whole repo on every push/deploy/testcheck, this nifty little utility checks only files changed with master.

Speedup is huge:

before:
yarn prettify:check  24.34s user 1.17s system 155% cpu 16.421 total

after:
yarn prettify:check  0.99s user 0.26s system 91% cpu 1.376 total

If we change a prettier rule then just use prettier as normal to reformat every file.

I'm not 100% certain how the logic works in every sort of git scenario, but testing with changes on master and on various branches it behaved as I would expect.

This change also upgrades Prettier to 2.0. I didn't want to change our style rules alongside this change, so updated our prettier config to keep existing styles for now.