@shaahmad 2020-08-27
Given a workflow of first pushing to master, then deploying to live, wouldn't this prevent prettier from running? Because your working tree will now resemble master, prettier will consider all files unchanged, and the unpretty files will be pushed to live.