@breck7 2020-08-27
> Given a workflow of first pushing to master, then deploying to live, wouldn't this prevent prettier from running? Because your working tree will now resemble master, prettier will consider all files unchanged, and the unpretty files will be pushed to live.

Yes. Good catch. So you could get away with pushing ugly files directly to master. I think that's a very good argument in favor of @MarcelGerber prettier GitHub action which could auto resolve those scenarios.