@danielgavrilov 2020-08-27
Awesome! If this is so quick now, we could run it in a git post-commit hook?