@breck7 2020-08-27
Pre-commit is a good idea. Here it is as a precommit hook: https://github.com/owid/owid-grapher/tree/husky

It works easy enough, but I'm not sure I'm crazy about it because it adds half a second of overhead to a commit (time output below). I'm not sure if it would get messy during a rebase or merge. Maybe a pre-push check? Or maybe just a precommit hook if on the master branch (but again, not sure how that will interfere with merges)?

In general I'm sort of not worried about the occasional ugly commit getting into master and getting prettified later. Right now I can make a commit to master on Git and not run the checks, so covering every route even today seems tough.

I think having a GitHub action on master might be the way to go. Then we can conceptualize it as just another dev who is very strict about prettify and anytime you commit to master will check and post a patch if it wasn't run.

Precommit time:

[husky 873c6b243] quick commit
 2 files changed, 2 insertions(+), 1 deletion(-)
 create mode 100644 wham.js
git commit -m 'quick commit'  0.04s user 0.09s system 68% cpu 0.184 total


🔍  Finding changed files since git revision d1500fc07.
🎯  Found 1 changed file.
✅  Everything is awesome!
[husky a9e6b3646] quick commit
 1 file changed, 1 insertion(+)
 create mode 100644 wham.js
git commit -m 'quick commit'  0.66s user 0.34s system 57% cpu 1.729 total